fix exo attempts label higher than on website cause it doesnt only take high sink stats, fix first mage is exo mage history hasn't changed, fix sink in mage resolve (integer) - fixes ini rune not used if 1 sink - only if 2, relevant min maxes

+++ Inkybot/Services/ExoItemMageResolve.cs
@@ -8,9 +8,9 @@ namespace Inkybot.Services
 {
     internal class ExoItemMageResolve : PrioritizedItemMageResolve
     {
-        public readonly float Sink;
+        public readonly int Sink;
         
-        public ExoItemMageResolve(MageConfig config, Item item, float sink) : base(config, item) =>
+        public ExoItemMageResolve(MageConfig config, Item item, int sink) : base(config, item) =>
             Sink = sink;
 
         protected override IEnumerable<ItemMage> PotentialMages() {
